requirements.txt
discord.py
dnspython
motor
requests
aiohttp
EZPaginator
Pillow
mathterpreter
imageio
pymongo
